Phase 4f: move firmware signing-key logic from Devices to Accounts
Continues relocating non-device logic out of the Devices context. Built on
Phase 4e (cleanup/phase-4e-firmware-delta-resolution).

Moves the two org-key functions out of NervesHub.Devices into NervesHub.Accounts,
which owns the OrgKey schema:
- fetch_firmware_signing_keys/1 (queries OrgKey by device)
- maybe_copy_firmware_keys/2 (copies a firmware's signing keys into the target
  org when a device moves orgs; now calls the local create_org_key/1)

Both operated purely on OrgKey/Org. The now-unused OrgKey alias is dropped from
Devices. Callers repointed: Devices.move (internal) and DeviceLink now call
Accounts.

Verification: mix format (bare --check-formatted, exit 0), credo, dialyzer
(0 unskipped), full test suite (1396 passing).
